requirements.txt
requests
tabulate
openai
python-dotenv
ratelimit
praw
textblob
colorama
ascii_art
